requirements.txt
# Core dependencies
numpy>=1.21.0,<2.0.0
scipy>=1.7.0
opencv-python>=4.5.0
scikit-image>=0.19.0
scikit-learn>=1.0.0
Pillow>=9.0.0

# OCR
pytesseract>=0.3.8

# Optimization
optuna>=3.0.0
hyperopt>=0.2.7

# Configuration
PyYAML>=6.0
python-dotenv>=0.19.0

# CLI
click>=8.0.0
tqdm>=4.62.0
rich>=12.0.0

# Data handling
pandas>=1.3.0
h5py>=3.6.0

# Visualization
matplotlib>=3.4.0
seaborn>=0.11.0
plotly>=5.0.0

# Testing
pytest>=7.0.0
pytest-cov>=3.0.0
pytest-mock>=3.6.0

# Type checking
mypy>=0.950
types-PyYAML
types-requests

# Code quality
black>=22.0.0
flake8>=4.0.0
isort>=5.10.0

# Documentation
sphinx>=4.5.0
sphinx-rtd-theme>=1.0.0

# Optional GPU acceleration
cupy-cuda11x>=10.0.0
numba>=0.55.0         

# Web API (optional)
fastapi>=0.85.0
uvicorn[standard]>=0.18.0
python-multipart>=0.0.5

# Dev requirements
ipython>=8.0.0
jupyter>=1.0.0
jupyterlab>=3.4.0
notebook>=6.4.0
ipywidgets>=7.7.0

# Profiling
line-profiler>=3.5.0
memory-profiler>=0.60.0
py-spy>=0.3.0